refactor: pull Channel --> Stream adapater into its own module
