+++ query/src/exec.rs
@@ -7,6 +7,7 @@ pub mod field;
 pub mod fieldlist;
 mod schema_pivot;
 pub mod seriesset;
+pub mod stream;
 pub mod stringset;
 mod task;
 pub use context::{DEFAULT_CATALOG, DEFAULT_SCHEMA};
